feat: add `typedarray.prototype.slice` (#81)

+++ packages/tools/shared/src/typed-array-prototype.ts
@@ -1,1 +1,1 @@
-export const TypedArrayPrototype = Object.getPrototypeOf(Uint8Array.prototype);
+export const TypedArrayPrototype = Object.getPrototypeOf(Uint8Array.prototype) as Uint8Array;
